pyproject.toml
[project]
name = "the-depositum"
version = "0.1.0"
description = "A reproducible data engineering pipeline for extracting, processing, and formatting public domain Catholic texts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Michael Link"}
]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "striprtf>=0.0.26",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "certifi>=2024.2.2",
    "urllib3>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

